4-scan_ai-Manus/clean_project/requirements.txt
# متطلبات Python لنظام Gaara Scan AI
# Python Requirements for Gaara Scan AI System

# ==========================================
# إطار العمل الأساسي
# Core Framework
# ==========================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ==========================================
# قاعدة البيانات
# Database
# ==========================================
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# ==========================================
# Redis والتخزين المؤقت
# Redis and Caching
# ==========================================
redis==5.0.1
aioredis==2.0.1

# ==========================================
# الأمان والمصادقة
# Security and Authentication
# ==========================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# ==========================================
# معالجة الصور
# Image Processing
# ==========================================
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.4
scikit-image==0.22.0

# ==========================================
# الذكاء الاصطناعي والتعلم الآلي
# AI and Machine Learning
# ==========================================
torch==2.1.1
torchvision==0.16.1
tensorflow==2.15.0
scikit-learn==1.3.2
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0

# ==========================================
# معالجة الملفات
# File Processing
# ==========================================
aiofiles==23.2.1
python-magic==0.4.27
openpyxl==3.1.2

# ==========================================
# HTTP والشبكات
# HTTP and Networking
# ==========================================
httpx==0.25.2
requests==2.31.0
websockets==12.0

# ==========================================
# المراقبة والسجلات
# Monitoring and Logging
# ==========================================
prometheus-client==0.19.0
structlog==23.2.0
psutil==5.9.6

# ==========================================
# التاريخ والوقت
# Date and Time
# ==========================================
python-dateutil==2.8.2
pytz==2023.3

# ==========================================
# التطوير والاختبار
# Development and Testing
# ==========================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# ==========================================
# أدوات مساعدة
# Utilities
# ==========================================
click==8.1.7
rich==13.7.0
typer==0.9.0
tqdm==4.66.1

# ==========================================
# تكامل السحابة
# Cloud Integration
# ==========================================
boto3==1.34.0
azure-storage-blob==12.19.0
google-cloud-storage==2.10.0

# ==========================================
# قواعد البيانات الإضافية
# Additional Databases
# ==========================================
elasticsearch==8.11.0
pymongo==4.6.0

# ==========================================
# معالجة النصوص
# Text Processing
# ==========================================
nltk==3.8.1
spacy==3.7.2

# ==========================================
# تصدير التقارير
# Report Export
# ==========================================
reportlab==4.0.7
fpdf2==2.7.6
weasyprint==60.2

# ==========================================
# إدارة التكوين
# Configuration Management
# ==========================================
python-dotenv==1.0.0
configparser==6.0.0

# ==========================================
# أمان إضافي
# Additional Security
# ==========================================
cryptography==41.0.8
pyotp==2.9.0

# ==========================================
# واجهات برمجة التطبيقات
# API Interfaces
# ==========================================
celery==5.3.4
kombu==5.3.4

# ==========================================
# تحسين الأداء
# Performance Optimization
# ==========================================
cachetools==5.3.2
lru-dict==1.3.0

